Move FilePerm function from cli to headscale
